Bundle the Project Gemini manifesto
Version 0.16.0 of the Project Gemini Speculative Specification is now
included with PSGemini, as conceptual help.  Users who might have
downloaded this module without knowing what Gemini is (not sure who
would) can run this command to learn everything about it and gemtext:

    PS C:\> Get-Help about_ProjectGemini

Because Solderpunk wrote it in British English, we have our first
localization.  I "Americanized" it for the regular en-US documents.

Thanks for nytpu on Libera Chat for unintentionally inspiring me to
include this with the module.  If any of you are looking for a UNIX-
style man page, look up his gemtext.5.  It's quite good.