Fix link parsing bug requiring space after `=>`.
Previous versions of this module required whitespace after the `=>`
delimiter that started a gemtext link line.  Failure to include a
space would cause the line to be parsed incorrectly.  As it turns
out, it was always perfectly acceptable __not__ to include a space
between `=>` and the URL.

For example:

```
=>https://geminiprotocol.net Gemini Protocol site
=> gemini://geminiprotocol.net Gemini Protocol site over Gemini
```

The first line would have been incorrectly parsed, while the next
would have been parsed correctly.

In a news update on August 28, 2024, this ambiguity was called out
and clarified in version 0.24.1 of the Gemini Protocol specification.
Thus, we're going to call this commit a bug fix that fixes a bug
in the link parsing code.