Change TOFU model to use hostname:port, per spec.
The new version (0.24.1) of the specification calls out that Gemini
clients must use hostname _and port_ when matching server certificates.
Thus, I've amended my code to do just that, on the off chance that
there is a Gemini capsule running on a non-standard port, and I've
made sure all cmdlet support specifying a port number, whether as part
of the URI or via a named parameter.

This necessitated changing the TOFU cache file format, so I invented
a versioning scheme (it matches the module version, super easy).
The cache file will be upgraded automatically when running the
Invoke-GeminiRequest cmdlet, though I'm sure that someone out there
will want to convert their TOFU cache files manually for some asinine
reason, so there's a cmdlet for that.  Expect it to be removed in a
future release if people use this for evil.